Image_Classification/requirements.txt
tensorflow
numpy
keras
pillow
matplotlib
opencv-python